Consertado erros no mapeamento, feito tela de Delete de Agência

+++ src/DEKL.CP.Domain/Contracts/Entities/IBankAgency.cs
@@ -2,7 +2,7 @@
 {
     public interface IBankAgency
     {
-        short Number { get; set; }
+        int Number { get; set; }
         string ManagerName { get; set; }
         string PhoneNumber { get; set; }
         string Email { get; set; }
